tweet routes same user validation

+++ Backend/routes/tweet.js
@@ -6,7 +6,8 @@ module.exports = ({
     getUserPosts,
     createPost,
     deletePost,
-    updatePost
+    updatePost,
+    getPostinfo
 }) => {
     /* GET users listing. */
 
@@ -39,10 +40,20 @@ module.exports = ({
     router.delete('/:id', (req, res) => {
         // usually id is in the params when you go into nested routes
         let post_id = req.params.id  
-        //add userid and post id check <----
-        deletePost(post_id)
-        .then(data => {
-            res.json(data);
+        let user_id = req.session.user_id   
+        //add userid and post user_id check <----
+        getPostinfo(post_id)
+        .then(postInfo=>{
+            if(postInfo.length > 0 && postInfo[0]['user_id']=== user_id ){
+                deletePost(post_id)
+                .then(data => {
+                res.json(data);
+                })
+            } else {
+                res.json({
+                    msg: 'Sorry not your post'
+                })
+            }
         })
         .catch(err => res.json({
             error: err.message
@@ -53,10 +64,20 @@ module.exports = ({
     router.put('/:id', (req, res) => {
         // usually id is in the params when you go into nested routes
         let post_id = req.params.id  
-        const { title , content } = req.body   
-        updatePost(post_id, title, content)
-        .then(data => {
-            res.json(data);
+        let user_id = req.session.user_id  
+        const { title , content } = req.body 
+        getPostinfo(post_id)
+        .then(postInfo=>{
+            if(postInfo.length > 0 && postInfo[0]['user_id']=== user_id ){  
+                updatePost(post_id, title, content)
+                .then(data => {
+                    res.json(data);
+                })
+            } else {
+                res.json({
+                    msg: 'Sorry not your post'
+                })
+            }
         })
         .catch(err => res.json({
             error: err.message
